Toolbar, Fixes, Preparations
Toolbar title now changes when switching between fragments.
Cleaned up some bits of code and you can not longer back out of the app when the nav drawer is open.
Started the transition to a fragment for inserting data. Not functional and need to figure out why my new app_insert_toolbar layout is NULL when setting the title.