Moved Insert Data
Insert data in its own activity. Functionality for adding new items is not yet there. Income and expense forms still need to be combined. This new activity has a custom toolbar that contains a back arrow to return to the home activity and has an info button within an overflow menu that brings up a dialog that will be filled out later.