Insert Layout Updates
Fields within the insert data activity now animate depending on what the type of item you have select.
The hints are also now correct depending on the type.
Refactored "type" to now be known as "title" within the application and throughout the code due to the conflict with the "type" of item.